+++ tests/TestMain.cpp
@@ -41,6 +41,7 @@ void runScripting();
 void runCrowd();
 void runMLP();
 void runUI();
+void runCompression();
 
 } // namespace meattest
 
@@ -64,6 +65,7 @@ int main() {
     meattest::runCrowd();
     meattest::runMLP();
     meattest::runUI();
+    meattest::runCompression();
 
     std::printf("\n%d checks, %d failures\n", meattest::g_checks, meattest::g_failures);
     return meattest::g_failures == 0 ? 0 : 1;